[WINDOWS] Create a windows warning free build.

+++ DHCPServer/DHCPServer.cpp
@@ -97,7 +97,7 @@ namespace Plugin {
 
         Core::ProxyType<Web::Response> result(PluginHost::Factories::Instance().Response());
         Core::TextSegmentIterator index(
-            Core::TextFragment(request.Path, _skipURL, request.Path.length() - _skipURL),
+            Core::TextFragment(request.Path, _skipURL, static_cast<uint16_t>(request.Path.length() - _skipURL)),
             false,
             '/');
