+++ DIALServer/DIALServer.cpp
@@ -248,7 +248,7 @@ namespace Plugin {
 
             ASSERT(_dialServiceImpl != nullptr);
 
-            _skipURL = service->WebPrefix().length();
+            _skipURL = static_cast<uint16_t>(service->WebPrefix().length());
 
             (*_deviceInfo) = _T("<?xml version=\"1.0\"?>")
                              _T("<root xmlns=\"urn:schemas-upnp-org:device-1-0\">")
@@ -375,7 +375,7 @@ namespace Plugin {
         ASSERT(_skipURL <= request.Path.length());
         // <GET> ../
         Core::ProxyType<Web::Response> result(PluginHost::Factories::Instance().Response());
-        Core::TextSegmentIterator index(Core::TextFragment(request.Path, _skipURL, request.Path.length() - _skipURL), false, '/');
+        Core::TextSegmentIterator index(Core::TextFragment(request.Path, _skipURL, static_cast<uint16_t>(request.Path.length()) - _skipURL), false, '/');
 
         // By default, we are in front of any element, jump onto the first element, which is if, there is something an empty slot.
         index.Next();
